Quick stab at using an abstract root class
